Spike re-work of vacancy field display
