+++ app/presenters/vacancy_template_decorator.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class VacancyTemplateDecorator < Draper::Decorator
+  delegate_all
+
+  include ReadableVacancy
+end
