+++ app/views/publishers/vacancies/_sections.html.slim
@@ -1,4 +1,4 @@
-= render "publishers/vacancies/vacancy_review_sections/job_details", vacancy: vacancy, current_organisation: current_organisation
+= render "publishers/vacancies/vacancy_review_sections/job_details", vacancy: vacancy, current_organisation: current_organisation, step_process: step_process
 
 - if vacancy.section_begun?(:about_the_role, step_process)
   = render "publishers/vacancies/vacancy_review_sections/about_the_role", vacancy: vacancy
