+++ app/views/publishers/vacancy_templates/_contract_type.html.slim
@@ -2,7 +2,7 @@
   - row.with_key
     = t("jobs.contract_type")
   - row.with_value
-    = vacancy_contract_type_with_duration(template)
+    = template.readable_contract_information
   - row.with_action text: t("buttons.change"),
     href: organisation_vacancy_template_build_path(template, :contract_information, "back_to_#{action_name}": "true"),
     visually_hidden_text: t("jobs.contract_type")
