+++ app/views/publishers/vacancy_templates/_job_role.html.slim
@@ -2,7 +2,7 @@
   - row.with_key
     = t("jobs.job_role")
   - row.with_value
-    = vacancy_readable_job_roles(template)
+    = template.readable_job_role
   - row.with_action text: t("buttons.change"),
     href: organisation_vacancy_template_build_path(template, :job_role, "back_to_#{action_name}": "true"),
     visually_hidden_text: t("publishers.vacancies.steps.job_role")
