+++ app/views/publishers/vacancy_templates/_key_stages.html.slim
@@ -2,7 +2,7 @@
   - row.with_key
     = t("jobs.key_stage", count: template.key_stages.count)
   - row.with_value
-    = vacancy_readable_key_stages(template)
+    = template.readable_key_stages
   - row.with_action text: t("buttons.change"),
     href: organisation_vacancy_template_build_path(template, :key_stages, "back_to_#{action_name}": "true"),
     visually_hidden_text: t("jobs.key_stage", count: template.key_stages.count)
