+++ app/views/publishers/vacancy_templates/_subjects.html.slim
@@ -2,7 +2,7 @@
   - row.with_key
     = t("jobs.subjects", count: template.subjects.count)
   - row.with_value
-    = template.subjects.any? ? vacancy_readable_subjects(template) : t("jobs.not_defined")
+    = template.subjects.any? ? template.readable_subjects : t("jobs.not_defined")
   - row.with_action text: t("buttons.change"),
     href: organisation_vacancy_template_build_path(template, :subjects, "back_to_#{action_name}": "true"),
     visually_hidden_text: t("jobs.subjects", count: template.subjects.count)
