+++ spec/presenters/vacancy_decorator_spec.rb
@@ -345,4 +345,75 @@ RSpec.describe VacancyDecorator do
       end
     end
   end
+
+  describe "#readable_key_stages" do
+    let(:vacancy) { build_stubbed(:vacancy, key_stages: %w[ks1 early_years]) }
+
+    it "joins them correctly" do
+      expect(decorated.readable_key_stages).to eq("Key stage 1, Early years")
+    end
+
+    context "when there are no subjects" do
+      let(:vacancy) { build_stubbed(:vacancy, key_stages: []) }
+
+      it "returns empty string" do
+        expect(decorated.readable_key_stages).to be_blank
+      end
+    end
+  end
+
+  describe "#readable_subjects" do
+    let(:vacancy) { build_stubbed(:vacancy, subjects: %w[Acrobatics Tapestry]) }
+
+    it "joins them correctly" do
+      expect(decorated.readable_subjects).to eq("Acrobatics, Tapestry")
+    end
+
+    context "when there are no subjects" do
+      let(:vacancy) { build_stubbed(:vacancy, subjects: []) }
+
+      it "returns empty string" do
+        expect(decorated.readable_subjects).to be_blank
+      end
+    end
+  end
+
+  describe "#readable_contract_information" do
+    let(:vacancy) do
+      build_stubbed(:vacancy, contract_type: contract_type,
+                              fixed_term_contract_duration: fixed_term_contract_duration,
+                              is_parental_leave_cover: is_parental_leave_cover)
+    end
+
+    context "when permanent" do
+      let(:contract_type) { :permanent }
+      let(:fixed_term_contract_duration) { "" }
+      let(:is_parental_leave_cover) { nil }
+
+      it "returns Permanent" do
+        expect(decorated.readable_contract_information).to eq "Permanent"
+      end
+    end
+
+    context "when fixed term" do
+      let(:contract_type) { :fixed_term }
+      let(:fixed_term_contract_duration) { "6 months" }
+
+      context "when is_parental_leave_cover is false" do
+        let(:is_parental_leave_cover) { false }
+
+        it "returns Fixed term (duration)" do
+          expect(decorated.readable_contract_information).to eq "Fixed term - 6 months"
+        end
+      end
+
+      context "when is_parental_leave_cover is true" do
+        let(:is_parental_leave_cover) { true }
+
+        it "returns Fixed term (duration)" do
+          expect(decorated.readable_contract_information).to eq "Fixed term - 6 months - Maternity or parental leave cover"
+        end
+      end
+    end
+  end
 end
